+++ src/components/FFlagIcon.vue
@@ -21,13 +21,16 @@ import { computed } from 'vue';
 
 import FSvgImage from './FSvgImage.vue';
 
-import type { CountryCode } from 'libphonenumber-js';
-
 export interface FFlagIconProps {
   /**
    * Code of the country, following the ISO 3166 standard
+   * @deprecated Use flagCode prop instead
    */
-  countryCode: CountryCode;
+  countryCode?: FlagCode;
+  /**
+   * Code of the flag, following the ISO 3166 standard + regional flags (see /src/flags.ts for full list)
+   */
+  flagCode: FlagCode;
   /**
    * Size of the flag icon
    */
@@ -35,7 +38,8 @@ export interface FFlagIconProps {
 }
 
 const props = withDefaults(defineProps<FFlagIconProps>(), {
-  countryCode: 'FR',
+  countryCode: undefined,
+  flagCode: 'FR',
   size: 24,
 });
 
@@ -52,7 +56,7 @@ const flagPaths = Object.keys(flagFiles);
 const currentFlagPath = computed(
   () =>
     flagPaths.find(path =>
-      new RegExp(`/${props.countryCode}.svg`).test(path)
+      new RegExp(`/${props.countryCode ?? props.flagCode}.svg`).test(path)
     ) ?? flagPaths[0]
 );
 </script>